requirements.txt
absl-py==0.11.0
astor==0.8.1
cached-property==1.5.2
click==7.1.2
cycler==0.10.0
Flask==1.1.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scipy==1.5.4
six==1.15.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
